tests: fs: Move testfs_utils to common directory
Move utils for file system tests into common directory to allow other
tests to use them.
