+++ tests/subsys/fs/littlefs/src/main.c
@@ -4,12 +4,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* Tests of functions in testfs_util */
-
 #include <string.h>
 #include <zephyr/ztest.h>
 #include "testfs_tests.h"
 #include "testfs_lfs.h"
-
 
 ZTEST_SUITE(littlefs, NULL, NULL, NULL, NULL, NULL);
