+++ tests/subsys/fs/littlefs/src/test_util.c
@@ -9,7 +9,7 @@
 #include <string.h>
 #include <zephyr/ztest.h>
 #include "testfs_tests.h"
-#include "testfs_util.h"
+#include "../../common/test_fs_util.h"
 
 #define MNT "/mnt"
 #define ELT1 "a"
